Fixing rebase moving around component that belongs to order
